pyproject.toml
[project]
name = "pr-opper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.21",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
    "uvicorn>=0.34.3",
]
